+++ src/components/input/input-decorator.tsx
@@ -8,7 +8,7 @@ interface Props {
   children: React.ReactNode;
 }
 
-export default class InputDecorator extends React.Component<Props> {
+export class InputDecorator extends React.Component<Props> {
   static contextType = FormContext;
   declare context: React.ContextType<typeof FormContext>;
